Change how options are handled
Options can be split into Core Options and other options.
Core Options are the ones that we -must- know in order to load the network.
These include anything that determines the shape of the network: image size,
latent vector size, and the selected network architectures.

When a result_dir is created, its Core Options are saved in params.json
All subsequent experiments on that result_dir will load and use those core options

Non-core options can be experiment-specific and might change with every run.