Add new sphere encoder for 32x32
This encoder maps all its outputs to the n-dimensional unit sphere.
So, the length (the Euclidian norm) of every vector it outputs will
always be 1. All the encoded points lie on a sphere, so their distance
from the center is the same. Only their direction matters.

Minor note: in order to make encoder32sphere work, we now have to pass
the batch size to the constructor for encoders. To generalize on this,
all network constructors now take a **kwargs argument, so if they need
batch size (or any other parameter) they can use it, and they can
ignore any they don't need.