Change active learning labeling
I don't yet have a way of learning open-set classification inside
the latent space. Right now I can only partition the latent space.

So, the system should be able to learn the decision boundaries
between class A and class B for each A,B. It might not be able
to learn 'dead space' zones within the latent space that do not
represent any class.

It's an open question whether there even should be a dead zone in
the latent space.