Switch UI result dir
A constant reminder to update this UI to work for any result_dir

The tricky thing there isn't just pointing it at a different dir,
it's ensuring that the pointed-to dir has trajectories.

Two ways of doing that. Work it into the always-up automation, or
add some special button to enqueue a batch of generate jobs