Change name of joint parameter again
In Argparse, you can specify an argument type and a default value.
But beware: if you use type=bool with default=True, the argument
will actually always be set to True, even if it's set to "False".

The trick is that "type" is actually a function that should take a
single parameter (a string). Not clear from the interface/docs.