Squeeze encoder output the right way
Using .squeeze() resizes the input array so that any dimensions of
size 1 are removed, "squeezed" away. That's often what you want when
you have eg. a batch_size x 128 x 1 x 1 convolutional output and
you want it to go into a fully connected layer

But what if the batch size is 1? Oops, then the batch dimension gets
squeezed away as well. This is a recurring problem for me with squeeze,
and so I recommend always specifying which axis to squeeze on eg.

.squeeze(2)