New mixed positive/negative training
This is great stuff, now you can have all-positive or all-negative
labels, and the system still works. Also you don't have to mess with
the learning rate or anything based on the positive/negative balance

Each batch can now be a mix of positive and negative, and the loss
functions for each kind mask out the other kind.