Implement multi-class discriminator
In a "Feature matching GAN" the discriminator predicts among K+1
classes where the last class is the "fake" one.

Here, we predict among K classes, but we also use the layer just
behind the softmax- we apply a hinge loss to it, and use it like
a one-versus-all classifier for each class.